symlinkPackages + createBinaryLink = src/utils/symlink
